Add basic tests to cmake and github action
  Also fixed a bug in src/util/linalg_wrapper.f90, the optional
  argument `info` was used in place of `wrk_info` without checking
  if it is present.

  This wasn't throwing an error because all calls to the subroutine
  included this optional argument, but a bug occured later on the
  line containing `info /= 0` which actually changed the value of
  `info` for some compiler/flag combinations due to some conversion
  between int32 and int64. The current change doesn't actually fix
  the underlying issue (the wrong integer kind is passed to the
  LAPACK subroutine), but the `info = wrk_info` line correctly
  converts the integer kind anyways.
