+++ .github/workflows/ci.yml
@@ -4,7 +4,7 @@ on:
   push:
 
 jobs:
-  build:
+  build-test:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -12,4 +12,12 @@ jobs:
         run: cmake --preset release
       - name: Build
         run: cmake --build --preset release
-
+      - name: Setup Python
+        uses: actions/setup-python@v5
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install numpy
+      - name: Run ZagHop Tests
+        working-directory: ./build/release
+        run: ctest
